Conflicts:
	org.lamport.tla.toolbox.product.product/TLAToolbox.target
	pom.xml
	tlatools/META-INF/MANIFEST.MF

+++ tlatools/src/tlc2/tool/ModelChecker.java
@@ -22,6 +22,7 @@ import tlc2.tool.queue.IStateQueue;
 import tlc2.util.IdThread;
 import tlc2.util.LongVec;
 import tlc2.util.ObjLongTable;
+import tlc2.util.statistics.BucketStatistics;
 import tlc2.value.Value;
 import util.DebugPrinter;
 import util.FileUtil;
@@ -38,6 +39,12 @@ import util.UniqueString;
 // the name resolver and support for the external specification object has been added
 public class ModelChecker extends AbstractChecker
 {
+
+	/**
+	 * If the state/ dir should be cleaned up after a successful model run
+	 */
+	private static final boolean VETO_CLEANUP = Boolean.getBoolean(ModelChecker.class.getName() + ".vetoCleanup");
+
     public FPSet theFPSet; // the set of reachable states (SZ: note the type)
     public IStateQueue theStateQueue; // the state queue
     public TLCTrace trace; // the trace file
@@ -93,14 +100,6 @@ public class ModelChecker extends AbstractChecker
         
         // needed to calculate state/minute in final progress report
         final long startTime = System.currentTimeMillis();
-
-        // Initialization for liveness checking:
-        if (this.checkLiveness)
-        {
-            report("initializing liveness checking");
-            LiveCheck.init(this.tool, this.actions, this.metadir);
-            report("liveness checking initialized");
-        }
 
         boolean recovered = this.recover();
         if (!recovered)
@@ -203,9 +202,10 @@ public class ModelChecker extends AbstractChecker
                 // Always check liveness properties at the end:
                 if (this.checkLiveness)
                 {
-                    MP.printMessage(EC.TLC_CHECKING_TEMPORAL_PROPS, "complete");
+					MP.printMessage(EC.TLC_CHECKING_TEMPORAL_PROPS,
+							new String[] { "complete", Long.toString(this.theFPSet.size()) });
                     report("checking liveness");
-                    success = LiveCheck.check();
+                    success = liveCheck.finalCheck();
                     report("liveness check complete");
                     if (!success)
                     {
@@ -238,7 +238,21 @@ public class ModelChecker extends AbstractChecker
             MP.printError(EC.GENERAL, e);  // LL changed call 7 April 2012
         } finally
         {
-            this.printSummary(success, startTime);
+        	
+        	this.printSummary(success, startTime);
+
+        	if (this.checkLiveness) {
+				if (LIVENESS_STATS) {
+					// Reclaim memory for in-degree calculation
+					System.gc();
+
+					MP.printStats(liveCheck
+							.calculateInDegreeDiskGraphs(new BucketStatistics("Histogram vertex in-degree",
+									LiveCheck.class.getPackage().getName(), "DiskGraphsInDegree")),
+							liveCheck.getOutDegreeStatistics());
+				}
+        	}
+
             this.cleanup(success);
         }
 
@@ -319,7 +333,7 @@ public class ModelChecker extends AbstractChecker
                         // build behavior graph for liveness checking
                         if (this.checkLiveness)
                         {
-                            LiveCheck.addInitState(curState, fp);
+                            liveCheck.addInitState(curState, fp);
                         }
                     }
                 }
@@ -593,7 +607,7 @@ public class ModelChecker extends AbstractChecker
                 long curStateFP = curState.fingerPrint();
                 liveNextStates.addElement(curState);
                 liveNextFPs.addElement(curStateFP);
-                LiveCheck.addNextState(curState, curStateFP, liveNextStates, liveNextFPs);
+                liveCheck.addNextState(curState, curStateFP, liveNextStates, liveNextFPs);
             }
             return false;
         } catch (Throwable e)
@@ -638,8 +652,8 @@ public class ModelChecker extends AbstractChecker
             boolean doCheck = this.checkLiveness && (stateNum >= nextLiveCheck);
             if (doCheck)
             {
-                MP.printMessage(EC.TLC_CHECKING_TEMPORAL_PROPS, "current");
-                if (!LiveCheck.check())
+				MP.printMessage(EC.TLC_CHECKING_TEMPORAL_PROPS, new String[] { "current", Long.toString(stateNum) });
+                if (!liveCheck.check())
                     return false;
                 nextLiveCheck = (stateNum <= 640000) ? stateNum * 2 : stateNum + 640000;
             }
@@ -655,7 +669,7 @@ public class ModelChecker extends AbstractChecker
             UniqueString.internTbl.beginChkpt(this.metadir);
             if (this.checkLiveness)
             {
-                LiveCheck.beginChkpt();
+                liveCheck.beginChkpt();
             }
             // commit checkpoint:
             this.theStateQueue.commitChkpt();
@@ -664,7 +678,7 @@ public class ModelChecker extends AbstractChecker
             UniqueString.internTbl.commitChkpt(this.metadir);
             if (this.checkLiveness)
             {
-                LiveCheck.commitChkpt();
+                liveCheck.commitChkpt();
             }
             MP.printMessage(EC.TLC_CHECKPOINT_END);
         }
@@ -683,7 +697,7 @@ public class ModelChecker extends AbstractChecker
             this.theFPSet.recover();
             if (this.checkLiveness)
             {
-                LiveCheck.recover();
+                liveCheck.recover();
             }
             MP.printMessage(EC.TLC_CHECKPOINT_RECOVER_END, new String[] { String.valueOf(this.theFPSet.size()),
                     String.valueOf(this.theStateQueue.size()) });
@@ -698,11 +712,13 @@ public class ModelChecker extends AbstractChecker
         this.theFPSet.close();
         this.trace.close();
         if (this.checkLiveness)
-            LiveCheck.close();
+            liveCheck.close();
         if (this.allStateWriter != null)
             this.allStateWriter.close();
-        FileUtil.deleteDir(this.metadir, success);
-    }
+        	if (!VETO_CLEANUP) {
+        		FileUtil.deleteDir(this.metadir, success);
+        	}
+		}
 
     public final void printSummary(boolean success, final long startTime) throws IOException
     {
@@ -1025,15 +1041,6 @@ public class ModelChecker extends AbstractChecker
             // }
             this.wait(TLCGlobals.progressInterval);
         }
-    }
-
-    /**
-     * Debugging support
-     * @param message
-     */
-    private void report(String message)
-    {
-        DebugPrinter.print(message);
     }
 
     /**
